feat: 添加 dropByCacheKey 用于清除缓存
