+++ packages/keep-alive/src/utils/getModelContent.tsx
@@ -1,31 +1,12 @@
-export default (absTmpPath: string) => `
-import React from 'react';
-import { routes } from '${absTmpPath}/core/routes';
-let BasicLayoutInstance: BasicLayout;
-const getKeepAliveViewMap = (routeList:any[],aliveList:any[])=>{
-  let keepAliveMap = {};
-  function find(routess: any[], list:any[]) {
-    if(!routess|| !list ){
-      return routess;
-    }
-    return routess.map(element => {
-      if (!Array.isArray(element.routes)&&list.includes(element.path)) {
-        element.recreateTimes = 0;
-        keepAliveMap[element.path] = element;
-      }else{
-        element.routes = find(element.routes,aliveList);
-      }
-      return element;
-    });
-  }
-  find(routeList,aliveList)
-  return keepAliveMap;
+export default () => `
+interface LayoutInstanceProps {
+  alivePathnames:string[],
+  keepAliveViewMap:{}
 }
-
-export function dropByCacheKey(pathname: string) {
-  console.log('dropByCacheKey',pathname)
-  if (BasicLayoutInstance) {
-    const { alivePathnames, keepAliveViewMap } = BasicLayoutInstance;
+let LayoutInstance:LayoutInstanceProps;
+function dropByCacheKey(pathname: string) {
+  if (LayoutInstance) {
+    const { alivePathnames, keepAliveViewMap } = LayoutInstance;
     const index = alivePathnames.findIndex(item => item === pathname);
     if (index !== -1) {
       alivePathnames.splice(index, 1);
@@ -34,73 +15,12 @@ export function dropByCacheKey(pathname: string) {
     }
   }
 }
-
-interface PageProps {
-  location: {
-    pathname: string;
-  };
+const setLayoutInstance = (value:any)=>{
+  LayoutInstance=value
 }
-export default class BasicLayout extends React.PureComponent<PageProps> {
-  constructor(props: any) {
-    super(props);
-    this.keepAliveViewMap = getKeepAliveViewMap(routes,props.keepalive);
-  }
-  componentDidMount() {
-    BasicLayoutInstance = this;
-  }
+const getLayoutInstance = ()=>LayoutInstance;
 
-  keepAliveViewMap = {};
-
-  alivePathnames: string[] = [];
-
-  render() {
-    const {
-      location: { pathname },
-    } = this.props;
-    const showKeepAlive = !!this.keepAliveViewMap[pathname];
-    if (showKeepAlive) {
-      const index = this.alivePathnames.findIndex(
-        tPathname => tPathname === pathname,
-      );
-      if (index === -1) {
-        this.alivePathnames.push(pathname);
-      }
-    }
-    const abc = (
-      <>
-        <div
-          style={{ position: 'relative' }}
-          hidden={!showKeepAlive}
-        >
-          {this.alivePathnames.map(curPathname => {
-            const View = this.keepAliveViewMap[curPathname].component;
-            return View ? (
-              <div
-                id={\`BasicLayout-\${curPathname}\`}
-                key={
-                  curPathname + this.keepAliveViewMap[curPathname].recreateTimes
-                }
-                style={{
-                  position: 'absolute',
-                  left: 0,
-                  top: 0,
-                  right: 0,
-                  bottom: 0,
-                }}
-                hidden={curPathname !== pathname}
-              >
-                <View {...this.props} />
-              </div>
-            ) : null;
-          })}
-        </div>
-        <div hidden={showKeepAlive}>
-          {!showKeepAlive && this.props.children}
-        </div>
-      </>
-    )
-    return abc;
-  }
+export {
+  setLayoutInstance,getLayoutInstance,dropByCacheKey
 }
-
 `;
